# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a docs for TodoList.",
        "title": "TodoList API",
        "contact": {
            "name": "ChengXiao",
            "email": "cxxlxx0@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/register": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Register",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get Jwt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.JwtResponse"
                        }
                    },
                    "401": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/group/": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Paged all groups",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Create one  group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group name",
                        "name": "group_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
        "/group/{group}": {
            "delete": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Delete one group by groupID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group hashID",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Rename group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group hashID",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "new group name",
                        "name": "new_group_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "FAIL",
                        "schema": {
                            "$ref": "#/definitions/serializers.CommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serializers.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "serializers.JwtResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expire_time": {
                    "type": "integer"
                }
            }
        }
    }
}